+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/AbstractJavaNode.java
@@ -5,6 +5,7 @@
 package net.sourceforge.pmd.lang.java.ast;
 
 import org.checkerframework.checker.nullness.qual.NonNull;
+import org.checkerframework.checker.nullness.qual.Nullable;
 
 import net.sourceforge.pmd.lang.ast.AstVisitor;
 import net.sourceforge.pmd.lang.ast.impl.javacc.AbstractJjtreeNode;
@@ -17,11 +18,19 @@ abstract class AbstractJavaNode extends AbstractJjtreeNode<AbstractJavaNode, Jav
 
     private Scope scope;
     protected JSymbolTable symbolTable;
-    private Comment comment;
     private ASTCompilationUnit root;
 
     AbstractJavaNode(int id) {
         super(id);
+    }
+
+    /**
+     * Temporary hack so that classes and methods are reported on their
+     * identifier token and not the first annotation. Changes about text
+     * documents make that more general, in a future PR.
+     */
+    protected @Nullable JavaccToken getPreferredReportLocation() {
+        return null;
     }
 
 
@@ -113,15 +122,6 @@ abstract class AbstractJavaNode extends AbstractJjtreeNode<AbstractJavaNode, Jav
         this.scope = scope;
     }
 
-    void comment(Comment theComment) {
-        comment = theComment;
-    }
-
-
-    @Override
-    public Comment comment() {
-        return comment;
-    }
 
     @Override
     @NonNull