+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/internal/JavaRuleViolationFactory.java
@@ -15,6 +15,8 @@ import net.sourceforge.pmd.Rule;
 import net.sourceforge.pmd.RuleViolation;
 import net.sourceforge.pmd.ViolationSuppressor;
 import net.sourceforge.pmd.lang.ast.Node;
+import net.sourceforge.pmd.lang.ast.impl.javacc.JavaccToken;
+import net.sourceforge.pmd.lang.java.ast.InternalApiBridge;
 import net.sourceforge.pmd.lang.java.ast.JavaNode;
 import net.sourceforge.pmd.lang.java.rule.JavaRuleViolation;
 import net.sourceforge.pmd.lang.rule.RuleViolationFactory;
@@ -49,7 +51,13 @@ public final class JavaRuleViolationFactory extends DefaultRuleViolationFactory 
 
     @Override
     public RuleViolation createViolation(Rule rule, @NonNull Node location, @NonNull String formattedMessage) {
-        return new JavaRuleViolation(rule, (JavaNode) location, formattedMessage);
+        JavaNode javaNode = (JavaNode) location;
+        JavaRuleViolation violation = new JavaRuleViolation(rule, javaNode, formattedMessage);
+        JavaccToken preferredLoc = InternalApiBridge.getReportLocation(javaNode);
+        if (preferredLoc != null) {
+            violation.setLines(preferredLoc.getBeginLine(), preferredLoc.getEndLine());
+        }
+        return violation;
     }
 
 }